feat(wffi): give a try to support ffi for Windows
As a vim native search plugin, it's not fair to use a hack way to get
position list in Windows.
hlslens will use FFI if Neovim version is 0.6 or later for Windows.

Note: It takes me nearly half a month to support FFI for *unix
stability, but I don’t use Windows, so I’m not sure if this commit will
cause random crashes. If you are a Windows user and encounter this
issue, please open an issue, and if I can't resolve it, I will revert
this commit.